requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
PyMySQL
passlib[bcrypt]
cryptography
python-jose[cryptography]
python-dotenv
pyjwt
fastapi-mail

